refactor(columnar): split ColumnarMapEquation.cpp into focused translation units
Pure move, no behaviour change. The file had grown to 4284 lines holding four
unrelated concerns, so the algorithm was buried under environment-variable knobs
and the objective arithmetic was interleaved with the search that consumes it.

  ColumnarTuning.h/.cpp      285  env-var A/B knobs, hsplit counters, the
                                  partial-seed policy, PhaseTimer
  ColumnarObjective.h        190  the base map equation as arithmetic: the move
                                  deltas (plain/tele, hoisted), the L* codebook
                                  terms, the calibration constants
  ColumnarParityCheck.cpp    233  ColumnarMapEquation, the read-only tree mirror
                                  behind --columnar-check
  ColumnarCorrections.cpp    715  the five ColumnarCorrection implementations
  ColumnarMapEquation.cpp   2960  ColumnarTwoLevel: the search

Two constraints shaped it rather than a line count:

- The native build has no LTO, so the per-candidate arithmetic must stay
  inlinable into the move loop. ColumnarObjective.h is a header for that reason,
  and moveLoop keeps removeModuleTerms/addModuleTerms in its own translation
  unit. Measured: min-of-9 interleaved against the pre-split binary, powergrid
  -C +0.52% and science2001 -C +0.06% (the pre-split arm's own spread on
  powergrid is 0.257-0.348s), web-NotreDame -0.51%, air30k -1.75%,
  malaria -1.22%. No systematic direction.
- MapEquation.h already declares OldSideTerms at namespace scope for the
  object-oriented objectives, so the columnar kernels cannot go there. They live
  in namespace infomap::columnar, which consumers open with a single
  `using namespace columnar;` — no call site changes.

Verified: all 38 recorded benchmark configs (13 networks x {OO, -C, -C
--non-redundant}) reproduce their codelength, top-module count and level count
exactly; make test-native 31/31 at OPENMP=1, OPENMP=0 and
FEATURES="lossy-map-equation regularized-multilayer"; clang-format clean.
Both build systems discover src/**/*.cpp, so no build files change.
