+++ src/core/ColumnarMapEquation.cpp
@@ -33,653 +33,12 @@
 #include <cstdio>
 #endif
 
+#include "ColumnarObjective.h"
+#include "ColumnarTuning.h"
+
 namespace infomap {
 
-// Co-physical (correction-proposed) candidate mode, from env COL_COMERGE:
-//   off (default) | all | seeded.
-// Proposing co-physical merge candidates is redundant with the edge-based
-// candidate set on UNDIRECTED clustering (co-physical state nodes are 2-hop
-// connected there) but may matter on DIRECTED clustering, where co-physical
-// modules need not be edge-adjacent. Kept as a tuning knob for that case:
-//   0 = off, 1 = all move-loop phases + merge, 2 = seeded phases only.
-static int coMergeMode()
-{
-  static const int mode = [] {
-    const char* e = std::getenv("COL_COMERGE");
-    if (!e) return 0;
-    if (std::string(e) == "all") return 1;
-    if (std::string(e) == "seeded") return 2;
-    return 0;
-  }();
-  return mode;
-}
-
-// Hierarchical split operator (experimental, see splitLevelModules), from env
-// COL_HSPLIT:
-//   off/unset (default) | 1 | all  = every stack level
-//   leaf      = only k == 0 (split a leaf module)
-//   interior  = only k >= 1 (split a module of modules)
-//   cheap     = every level, but no fresh from-singletons leaf re-derivation
-//               (the one expensive piece source; F20 measured it at 3.17 s on
-//               malaria in the two-level operator)
-//   auto      = the measured policy: interior levels always (that is where the
-//               enter-flow up-build over-merges, on base networks too), the
-//               leaf level only when a module-move-capable correction is
-//               attached (where splitTopModules' gate does apply) and then
-//               only from the free piece sources
-//   winner    = the once-per-run policy: interior levels always, the leaf level
-//               only with a module-move-capable correction attached (as `auto`),
-//               but WITH the expensive from-singletons piece source, which is
-//               the discovery the once-per-run pass exists to pay for
-// Default off keeps the engine bit-identical to the pre-operator baseline.
-enum HSplitMode : std::uint8_t {
-  kHSplitOff = 0,
-  kHSplitAll = 1,
-  kHSplitLeaf = 2,
-  kHSplitInterior = 3,
-  kHSplitCheap = 4,
-  kHSplitAuto = 5,
-  kHSplitWinner = 6
-};
-static int hSplitParseMode(const char* envName, int whenUnset)
-{
-  const char* e = std::getenv(envName);
-  if (e == nullptr)
-    return whenUnset;
-  if (std::string(e) == "off")
-    return static_cast<int>(kHSplitOff);
-  const std::string v(e);
-  if (v == "1" || v == "all")
-    return static_cast<int>(kHSplitAll);
-  if (v == "leaf")
-    return static_cast<int>(kHSplitLeaf);
-  if (v == "interior")
-    return static_cast<int>(kHSplitInterior);
-  if (v == "cheap")
-    return static_cast<int>(kHSplitCheap);
-  if (v == "auto")
-    return static_cast<int>(kHSplitAuto);
-  if (v == "winner")
-    return static_cast<int>(kHSplitWinner);
-  return static_cast<int>(kHSplitOff);
-}
-// Per-trial variant: run the split interleave inside every trial's coarsening.
-static int hSplitMode()
-{
-  static const int mode = hSplitParseMode("COL_HSPLIT", static_cast<int>(kHSplitOff));
-  return mode;
-}
-// Once-per-run variant: run the split interleave ONCE on the winning trial's
-// hierarchy instead of inside every trial — the #889 maybeDeepRepairBest
-// pattern, which previously only fired for two-level-shaped winners.
-// Independent of COL_HSPLIT.
-//
-// ON BY DEFAULT: this is the shipped behaviour (malaria -0.371% mean over
-// seeds 123/234/345 for +5..11% CPU on a ~3s run; air30k and regularized
-// microscopic gains at +1.5..3%; every base network bit-identical at zero
-// cost, because the repair is gated on module-move corrections). Set
-// COL_HSPLIT_WINNER=off to disable it, or to any other mode name to override
-// the level policy.
-static int hSplitWinnerMode()
-{
-  static const int mode = hSplitParseMode("COL_HSPLIT_WINNER", static_cast<int>(kHSplitWinner));
-  return mode;
-}
-
-static bool hSplitLevelEnabled(int mode, int k)
-{
-  switch (mode) {
-  case kHSplitOff:
-    return false;
-  case kHSplitLeaf:
-    return k == 0;
-  case kHSplitInterior:
-    return k > 0;
-  default:
-    return true;
-  }
-}
-
-// Accepted/attempted split counters per stack level (COL_HSPLIT_STATS=1 prints
-// them at exit). Atomic so parallel trials stay countable.
-constexpr int kHSplitMaxLevels = 8;
-static std::atomic<long long> g_hSplitAttempts[kHSplitMaxLevels];
-static std::atomic<long long> g_hSplitAccepts[kHSplitMaxLevels];
-static std::atomic<long long> g_hSplitClocks[kHSplitMaxLevels]; // CPU clocks inside the operator
-// Cumulative RELATIVE codelength gain per level, in units of 1e-9, so the
-// accumulator can stay an integer atomic (gains are ~1e-6..1e-3 each).
-static std::atomic<long long> g_hSplitGainNano[kHSplitMaxLevels];
-// Sub-phase clocks inside the operator (diagnostic, COL_HSPLIT_STATS):
-// 0 = moveBase copy, 1 = piece aggregateLevel, 2 = seeded moveLoop,
-// 3 = stack rebuild (aggregations), 4 = hierarchicalCodelengthFromStack,
-// 5 = save/restore of levels+assignments, 6 = fresh sub-clustering.
-constexpr int kHSplitPhases = 7;
-static std::atomic<long long> g_hSplitPhase[kHSplitPhases];
-static const char* const kHSplitPhaseName[kHSplitPhases] = { "moveBaseCopy", "pieceAgg", "moveLoop", "rebuild", "evalStack", "saveRestore", "subCluster" };
-static bool hSplitPhaseTiming()
-{
-  static const bool on = std::getenv("COL_HSPLIT_PHASES") != nullptr;
-  return on;
-}
-namespace {
-  struct PhaseTimer {
-    std::clock_t t0;
-    int idx;
-    bool on;
-    explicit PhaseTimer(int i, bool enabled)
-        : t0(enabled ? std::clock() : 0), idx(i), on(enabled) {}
-    ~PhaseTimer()
-    {
-      if (on)
-        g_hSplitPhase[idx].fetch_add(static_cast<long long>(std::clock() - t0), std::memory_order_relaxed);
-    }
-  };
-
-  struct HSplitStatsDump {
-    ~HSplitStatsDump()
-    {
-      if (std::getenv("COL_HSPLIT_STATS") == nullptr)
-        return;
-      long long ta = 0, tc = 0;
-      double tt = 0.0, tg = 0.0;
-      for (int k = 0; k < kHSplitMaxLevels; ++k) {
-        const long long a = g_hSplitAttempts[k].load(std::memory_order_relaxed);
-        const long long c = g_hSplitAccepts[k].load(std::memory_order_relaxed);
-        const double t = static_cast<double>(g_hSplitClocks[k].load(std::memory_order_relaxed)) / CLOCKS_PER_SEC;
-        const double g = static_cast<double>(g_hSplitGainNano[k].load(std::memory_order_relaxed)) * 1e-9;
-        ta += a;
-        tc += c;
-        tt += t;
-        tg += g;
-        if (a != 0) {
-          const double perAtt = 100.0 * g / static_cast<double>(a);
-          std::fprintf(stderr, "[hsplit] level k=%d: %lld accepted / %lld attempted, %.2fs, gain %.4f%% (%.5f%%/att)\n", k, c, a, t, 100.0 * g, perAtt);
-        }
-      }
-      std::fprintf(stderr, "[hsplit] total: %lld accepted / %lld attempted, %.2fs, gain %.4f%%\n", tc, ta, tt, 100.0 * tg);
-      if (hSplitPhaseTiming())
-        for (int p = 0; p < kHSplitPhases; ++p) {
-          const double secs = static_cast<double>(g_hSplitPhase[p].load(std::memory_order_relaxed)) / CLOCKS_PER_SEC;
-          std::fprintf(stderr, "[hsplit] phase %-12s %.3fs\n", kHSplitPhaseName[p], secs);
-        }
-    }
-  };
-  HSplitStatsDump g_hSplitStatsDump;
-} // namespace
-
-// PARTIAL SEEDING of the within-grandparent layer refinements (the mechanism is
-// documented on buildPartialSeed).
-//
-// The shipped policy is (q = 0.40, bnd, re-refine only, every interior layer).
-// The env knobs are the A/B handles it was chosen with; COL_PARTSEED_Q=1
-// restores the from-singletons search bit-exactly.
-//
-//   COL_PARTSEED_Q=<0..1>  release fraction: the share of a grandparent's units
-//     the refine gets back as fresh singletons. q = 1 releases everything, i.e.
-//     the from-singletons default; q = 0 seeds everything, which is a fixpoint
-//     (the greedy loop reproduces its input and the gate reverts). The optimum
-//     sits on a broad plateau — web-NotreDame varies by < 0.03% over q in
-//     0.33-0.50 and powergrid by < 0.06% over 0.25-0.40 — and both degrade
-//     above 0.6, where too little is locked to be worth seeding at all.
-//   COL_PARTSEED_M=bnd|ex|inv|iex|rand   which units count as "loosest":
-//     bnd  share of the unit's link flow, inside the grandparent, that crosses
-//          its current module's boundary. The default: it is module-aware and
-//          well-defined on every flow model.
-//     ex   the unit's own leakiness, exit / (flow + exit). DEGENERATE on
-//          undirected first-order networks, where a leaf's exit flow equals its
-//          flow, so the ratio is 0.5 for every node and the "ranking" collapses
-//          to node order (measured: ex and iex are bit-identical on powergrid).
-//     inv / iex  controls for bnd / ex: same ranking, released from the tight
-//          end instead. Both must LOSE to the from-singletons default, or the
-//          effect is release volume rather than release targeting.
-//     rand release a uniformly random q (deterministic in m_seed): the other
-//          control, for "any perturbation would do".
-//   COL_PARTSEED_LEAF=1   restrict to the leaf layer (default: every interior
-//                         layer — the interior layers are where the sweeps that
-//                         re-refine actually live, and they are nearly free)
-//   COL_PARTSEED_ALWAYS=1 also partial-seed a layer's FIRST refine (default:
-//                         only re-refines, see partSeedResweepOnly)
-//   COL_PARTSEED_FLAT=1   apply it to a converged flat bottom instead of
-//                         skipping that layer (see m_bottomConverged)
-static constexpr double kPartSeedRelease = 0.40;
-enum class PartSeedMetric : std::uint8_t {
-  Boundary,
-  Exit,
-  InvBoundary,
-  InvExit,
-  Random
-};
-static double partSeedRelease()
-{
-  static const double q = [] {
-    const char* e = std::getenv("COL_PARTSEED_Q");
-    return e != nullptr ? std::atof(e) : kPartSeedRelease;
-  }();
-  return q;
-}
-static PartSeedMetric partSeedMetric()
-{
-  static const PartSeedMetric m = [] {
-    const char* e = std::getenv("COL_PARTSEED_M");
-    if (e == nullptr)
-      return PartSeedMetric::Boundary;
-    const std::string v(e);
-    if (v == "ex")
-      return PartSeedMetric::Exit;
-    if (v == "inv")
-      return PartSeedMetric::InvBoundary;
-    if (v == "iex")
-      return PartSeedMetric::InvExit;
-    if (v == "rand")
-      return PartSeedMetric::Random;
-    return PartSeedMetric::Boundary;
-  }();
-  return m;
-}
-static bool partSeedLeafOnly()
-{
-  static const bool on = std::getenv("COL_PARTSEED_LEAF") != nullptr;
-  return on;
-}
-static bool partSeedFlatBottom()
-{
-  static const bool on = std::getenv("COL_PARTSEED_FLAT") != nullptr;
-  return on;
-}
-// Partial-seed only a RE-refine of a layer, never its first from-singletons
-// derivation (COL_PARTSEED_ALWAYS=1 lifts this). Seeding damage is localised to
-// the first refine of a grandparent (F24 C), and before that first refine the
-// layer holds no partition worth locking — the build's greedy enter-flow
-// super-search made it, not a two-level optimize. Refine sweep 0 refines every
-// dirty layer, so m_refineSweep > 0 is exactly "this layer has been re-derived".
-// This is also what keeps the whole feature inert on the single-sweep searches
-// (`-F`, and any stack with one interior layer): they only ever refine once, so
-// there is nothing to re-refine and nothing changes.
-static bool partSeedResweepOnly()
-{
-  static const bool on = std::getenv("COL_PARTSEED_ALWAYS") == nullptr;
-  return on;
-}
-// Partial seeding applies to layer k? Off (q >= 1) is the from-singletons
-// default, which must stay bit-identical, so this is the single gate.
-static bool partSeedActive(int layer)
-{
-  const double q = partSeedRelease();
-  return q >= 0.0 && q < 1.0 && (layer == 0 || !partSeedLeafOnly());
-}
-// Deterministic 64-bit mix (splitmix64 finalizer): the PRNG for the `rand`
-// control, keyed so a run is reproducible from (m_seed, layer, unit).
-static inline unsigned long long partSeedMix(unsigned long long x)
-{
-  x += 0x9e3779b97f4a7c15ULL;
-  x = (x ^ (x >> 30)) * 0xbf58476d1ce4e5b9ULL;
-  x = (x ^ (x >> 27)) * 0x94d049bb133111ebULL;
-  return x ^ (x >> 31);
-}
-
-void ColumnarMapEquation::buildFromTree(const InfoNode& root, const std::vector<InfoNode*>& leafNodes, bool undirected)
-{
-  m_undirected = undirected;
-  const std::size_t nLeaves = leafNodes.size();
-
-  m_leafFlow.assign(nLeaves, 0.0);
-  std::unordered_map<const InfoNode*, int> leafId;
-  leafId.reserve(nLeaves * 2);
-  for (std::size_t i = 0; i < nLeaves; ++i) {
-    m_leafFlow[i] = leafNodes[i]->data.flow;
-    leafId[leafNodes[i]] = static_cast<int>(i);
-  }
-
-  // DFS over internal nodes (pre-order): assign module ids, depth, parent.
-  // A node is always popped after its parent, so the parent id is known.
-  m_modFlow.clear();
-  m_modEnter.clear();
-  m_modExit.clear();
-  m_modParent.clear();
-  m_modDepth.clear();
-  m_modLeafModule.clear();
-  std::vector<const InfoNode*> modNode;
-  std::unordered_map<const InfoNode*, int> modId;
-
-  std::vector<std::pair<const InfoNode*, int>> stack;
-  stack.emplace_back(&root, 0);
-  while (!stack.empty()) {
-    const InfoNode* node = stack.back().first;
-    const int depth = stack.back().second;
-    stack.pop_back();
-
-    const int id = static_cast<int>(m_modFlow.size());
-    modId[node] = id;
-    m_modFlow.push_back(0.0);
-    m_modEnter.push_back(0.0);
-    m_modExit.push_back(0.0);
-    m_modDepth.push_back(depth);
-
-    int par = -1;
-    if (node->parent != nullptr) {
-      auto it = modId.find(node->parent);
-      if (it != modId.end())
-        par = it->second;
-    }
-    m_modParent.push_back(par);
-    m_modLeafModule.push_back(node->firstChild != nullptr && node->firstChild->isLeaf() ? 1 : 0);
-    modNode.push_back(node);
-
-    for (const auto& child : *node) {
-      if (!child.isLeaf())
-        stack.emplace_back(&child, depth + 1);
-    }
-  }
-  const int nMod = static_cast<int>(m_modFlow.size());
-
-  // Children as CSR (second pass: all ids now assigned).
-  m_childStart.assign(nMod + 1, 0);
-  for (int m = 0; m < nMod; ++m) {
-    int cnt = 0;
-    for (const auto& child : *modNode[m]) {
-      (void)child;
-      ++cnt;
-    }
-    m_childStart[m + 1] = m_childStart[m] + cnt;
-  }
-  m_childList.assign(static_cast<std::size_t>(m_childStart[nMod]), 0);
-  for (int m = 0; m < nMod; ++m) {
-    int pos = m_childStart[m];
-    const bool leafMod = m_modLeafModule[m] != 0;
-    for (const auto& child : *modNode[m]) {
-      m_childList[pos++] = leafMod ? leafId.at(&child) : modId.at(&child);
-    }
-  }
-
-  // Leaf -> parent module id.
-  m_leafParentMod.assign(nLeaves, -1);
-  for (std::size_t i = 0; i < nLeaves; ++i) {
-    m_leafParentMod[i] = modId.at(leafNodes[i]->parent);
-  }
-
-  aggregate(leafNodes);
-}
-
-void ColumnarMapEquation::aggregate(const std::vector<InfoNode*>& leafNodes)
-{
-  const int nMod = static_cast<int>(m_modFlow.size());
-  std::fill(m_modFlow.begin(), m_modFlow.end(), 0.0);
-  std::fill(m_modEnter.begin(), m_modEnter.end(), 0.0);
-  std::fill(m_modExit.begin(), m_modExit.end(), 0.0);
-
-  // Module flow: deepest first so children are summed before their parents.
-  std::vector<int> order(nMod);
-  std::iota(order.begin(), order.end(), 0);
-  std::sort(order.begin(), order.end(), [this](int a, int b) { return m_modDepth[a] > m_modDepth[b]; });
-  for (int m : order) {
-    const bool leafMod = m_modLeafModule[m] != 0;
-    double f = 0.0;
-    for (int k = m_childStart[m]; k < m_childStart[m + 1]; ++k) {
-      const int c = m_childList[k];
-      f += leafMod ? m_leafFlow[c] : m_modFlow[c];
-    }
-    m_modFlow[m] = f;
-  }
-
-  // Enter/exit flow: walk each leaf out-edge from the two endpoints' modules up
-  // to their lowest common ancestor, matching aggregateFlowValuesFromLeafToRoot.
-  const std::size_t nLeaves = leafNodes.size();
-  std::unordered_map<const InfoNode*, int> leafId;
-  leafId.reserve(nLeaves * 2);
-  for (std::size_t i = 0; i < nLeaves; ++i)
-    leafId[leafNodes[i]] = static_cast<int>(i);
-
-  for (std::size_t i = 0; i < nLeaves; ++i) {
-    InfoNode* s = leafNodes[i];
-    for (InfoEdge* e : s->outEdges()) {
-      auto tIt = leafId.find(e->target);
-      if (tIt == leafId.end())
-        continue; // target not a leaf in this set (shouldn't happen)
-      int a = m_leafParentMod[i];
-      int b = m_leafParentMod[tIt->second];
-      if (a == b)
-        continue; // intra-module edge: no contribution to module enter/exit
-
-      const double f = e->data.flow;
-      const double half = f / 2.0;
-      while (m_modDepth[a] > m_modDepth[b]) {
-        if (m_undirected) {
-          m_modExit[a] += half;
-          m_modEnter[a] += half;
-        } else {
-          m_modExit[a] += f;
-        }
-        a = m_modParent[a];
-      }
-      while (m_modDepth[b] > m_modDepth[a]) {
-        if (m_undirected) {
-          m_modEnter[b] += half;
-          m_modExit[b] += half;
-        } else {
-          m_modEnter[b] += f;
-        }
-        b = m_modParent[b];
-      }
-      while (a != b) {
-        if (m_undirected) {
-          m_modExit[a] += half;
-          m_modEnter[a] += half;
-          m_modEnter[b] += half;
-          m_modExit[b] += half;
-        } else {
-          m_modExit[a] += f;
-          m_modEnter[b] += f;
-        }
-        a = m_modParent[a];
-        b = m_modParent[b];
-      }
-    }
-  }
-}
-
-double ColumnarMapEquation::moduleTerm(int m) const
-{
-  using infomath::plogp;
-  if (m_modLeafModule[m] != 0) {
-    // Module of leaf nodes: index codebook over child flow + module exit.
-    const double T = m_modFlow[m] + m_modExit[m];
-    if (T < 1e-16)
-      return 0.0;
-    double indexLength = 0.0;
-    for (int k = m_childStart[m]; k < m_childStart[m + 1]; ++k)
-      indexLength -= plogp(m_leafFlow[m_childList[k]] / T);
-    indexLength -= plogp(m_modExit[m] / T);
-    return indexLength * T;
-  }
-  // Module of modules: exit to coarser level or enter one of the children.
-  if (m_modFlow[m] < 1e-16)
-    return 0.0;
-  double sumEnter = 0.0;
-  double sumEnterLogEnter = 0.0;
-  for (int k = m_childStart[m]; k < m_childStart[m + 1]; ++k) {
-    const double en = m_modEnter[m_childList[k]];
-    sumEnter += en;
-    sumEnterLogEnter += plogp(en);
-  }
-  const double totalCodewordUse = m_modExit[m] + sumEnter;
-  return plogp(totalCodewordUse) - sumEnterLogEnter - plogp(m_modExit[m]);
-}
-
-double ColumnarMapEquation::hierarchicalCodelength() const
-{
-  double total = 0.0;
-  const int nMod = static_cast<int>(m_modFlow.size());
-  for (int m = 0; m < nMod; ++m)
-    total += moduleTerm(m);
-  return total;
-}
-
-unsigned int ColumnarMapEquation::numLevels() const
-{
-  int maxDepth = 0;
-  for (int d : m_modDepth)
-    maxDepth = std::max(maxDepth, d);
-  // Module depths run 0 (root) .. maxDepth (deepest leaf module); the leaves sit
-  // one level below, so the tree depth (matching maxTreeDepth) is maxDepth + 1.
-  return static_cast<unsigned int>(maxDepth + 1);
-}
-
-// ===================================================
-// ColumnarTwoLevel (M1b): columnar two-level optimizer
-// ===================================================
-
-namespace {
-  // Codelength calibration constants, mirroring the Config defaults the OO core
-  // uses (the columnar path is a measurement scaffold, not user-tunable).
-  constexpr double kMinImprovement = 1e-10;
-  constexpr double kMinSingleImprovement = 1e-16;
-  constexpr unsigned int kCoreLoopLimit = 10;
-
-  // Flat-first trials (#889): the cheap flat probe (full aggregation, no leaf
-  // fine-tune) must land within this relative margin of the fine-blocks
-  // up-build's post-build codelength for the trial to complete the expensive
-  // leaf-level flat pipeline. Measured across the benchmark set, every network
-  // where the flat search truly wins probes at est/build <= 1.000 (air30k
-  // 0.84, malaria 0.88, jazz 0.996, lazega 0.995) and every true flat-loser
-  // probes at >= 1.008 (science2001 1.008, netsci 1.034, powergrid 1.10,
-  // web-NotreDame 1.18) — the post-build refinement gains more than the flat
-  // completion does, so a generous margin only buys false positives. 0.5%
-  // sits in the middle of the observed gap.
-  constexpr double kFlatProbeMargin = 0.005;
-
-  // Exact port of MapEquation::getDeltaCodelengthOnMovingNode for the base
-  // objective: change in codelength from moving a unit out of its old module
-  // (aggregates old*) into a candidate module (aggregates new*). deltaOld/deltaNew
-  // are the summed enter+exit flow between the unit and the old/new module.
-  inline double deltaCodelengthMovingNode(double enterFlow, double enterFlow_log_enterFlow, double curEnter, double curExit, double curFlow, double oldEnter, double oldExit, double oldFlow, double newEnter, double newExit, double newFlow, double deltaOld, double deltaNew)
-  {
-    using infomath::plogp;
-    const double oldExitPlusFlow = oldExit + oldFlow;
-    const double newExitPlusFlow = newExit + newFlow;
-
-    const double delta_enter = plogp(enterFlow + deltaOld - deltaNew) - enterFlow_log_enterFlow;
-    const double delta_enter_log_enter = -plogp(oldEnter) - plogp(newEnter)
-        + plogp(oldEnter - curEnter + deltaOld) + plogp(newEnter + curEnter - deltaNew);
-    const double delta_exit_log_exit = -plogp(oldExit) - plogp(newExit)
-        + plogp(oldExit - curExit + deltaOld) + plogp(newExit + curExit - deltaNew);
-    const double delta_flow_log_flow = -plogp(oldExitPlusFlow) - plogp(newExitPlusFlow)
-        + plogp(oldExitPlusFlow - curExit - curFlow + deltaOld) + plogp(newExitPlusFlow + curExit + curFlow - deltaNew);
-
-    return delta_enter - delta_enter_log_enter - delta_exit_log_exit + delta_flow_log_flow;
-  }
-
-  // Old-module plogp terms of deltaCodelengthMovingNode: constant for every
-  // candidate the move loop probes for the same unit, so the caller computes
-  // them once per unit (6 of the 13 plogp calls per candidate hoist out; the
-  // per-candidate math below keeps the exact expression structure, so results
-  // stay bit-identical to the unhoisted form).
-  struct OldSideTerms {
-    double plogpOldEnter, plogpOldExit, plogpOldEF;
-    double plogpOldEnterAfter, plogpOldExitAfter, plogpOldEFAfter;
-  };
-
-  inline OldSideTerms hoistOldSide(double curEnter, double curExit, double curFlow, double oldEnter, double oldExit, double oldFlow, double deltaOld)
-  {
-    using infomath::plogp;
-    const double oldExitPlusFlow = oldExit + oldFlow;
-    return { plogp(oldEnter), plogp(oldExit), plogp(oldExitPlusFlow), plogp(oldEnter - curEnter + deltaOld), plogp(oldExit - curExit + deltaOld), plogp(oldExitPlusFlow - curExit - curFlow + deltaOld) };
-  }
-
-  inline double deltaCodelengthMovingNodeHoisted(double enterFlow, double enterFlow_log_enterFlow, double curEnter, double curExit, double curFlow, const OldSideTerms& o, double newEnter, double newExit, double newFlow, double deltaOld, double deltaNew)
-  {
-    using infomath::plogp;
-    const double newExitPlusFlow = newExit + newFlow;
-
-    const double delta_enter = plogp(enterFlow + deltaOld - deltaNew) - enterFlow_log_enterFlow;
-    const double delta_enter_log_enter = -o.plogpOldEnter - plogp(newEnter)
-        + o.plogpOldEnterAfter + plogp(newEnter + curEnter - deltaNew);
-    const double delta_exit_log_exit = -o.plogpOldExit - plogp(newExit)
-        + o.plogpOldExitAfter + plogp(newExit + curExit - deltaNew);
-    const double delta_flow_log_flow = -o.plogpOldEF - plogp(newExitPlusFlow)
-        + o.plogpOldEFAfter + plogp(newExitPlusFlow + curExit + curFlow - deltaNew);
-
-    return delta_enter - delta_enter_log_enter - delta_exit_log_exit + delta_flow_log_flow;
-  }
-
-  // Recorded-teleportation (tele-path) analogue of OldSideTerms/hoistOldSide:
-  // the six old-module (A-side) plogp terms plus the A-side shift of enterFlow
-  // are constant across every candidate the move loop probes for the same unit,
-  // so the caller hoists them once per unit visit. moduleTeleEnter/moduleTeleExit
-  // are inlined here as free lambdas (these helpers live in the class); the
-  // per-candidate expression structure below matches deltaCodelengthMovingNodeTele
-  // exactly, so results stay bit-identical to the unhoisted form.
-  struct TeleOldSideTerms {
-    double plogpEA0, plogpXA0, plogpFA0; // before the move
-    double plogpEA1, plogpXA1, plogpFA1; // after
-    double deltaEnterA; // eA1 - eA0, the A-side contribution to enterFlow1
-  };
-
-  inline TeleOldSideTerms hoistOldSideTele(double curEnter, double curExit, double curFlow, double tfu, double twu, double oldEnter, double oldExit, double oldFlow, double oldTeleFlow, double oldTeleWeight, double totalTeleFlow, double deltaOld)
-  {
-    using infomath::plogp;
-    const auto teleEnter = [&](double tf, double tw) { return (totalTeleFlow - tf) * tw; };
-    const auto teleExit = [&](double tf, double tw) { return tf * (1.0 - tw); };
-    const double eA0 = oldEnter + teleEnter(oldTeleFlow, oldTeleWeight);
-    const double xA0 = oldExit + teleExit(oldTeleFlow, oldTeleWeight);
-    const double fA0 = xA0 + oldFlow;
-    const double eA1 = (oldEnter - curEnter + deltaOld) + teleEnter(oldTeleFlow - tfu, oldTeleWeight - twu);
-    const double xA1 = (oldExit - curExit + deltaOld) + teleExit(oldTeleFlow - tfu, oldTeleWeight - twu);
-    const double fA1 = xA1 + (oldFlow - curFlow);
-    return { plogp(eA0), plogp(xA0), plogp(fA0), plogp(eA1), plogp(xA1), plogp(fA1), eA1 - eA0 };
-  }
-
-  inline double deltaCodelengthMovingNodeTeleHoisted(double enterFlow, double enterFlow_log_enterFlow, double curEnter, double curExit, double curFlow, double tfu, double twu, const TeleOldSideTerms& o, double newEnter, double newExit, double newFlow, double newTeleFlow, double newTeleWeight, double totalTeleFlow, double deltaNew)
-  {
-    using infomath::plogp;
-    const auto teleEnter = [&](double tf, double tw) { return (totalTeleFlow - tf) * tw; };
-    const auto teleExit = [&](double tf, double tw) { return tf * (1.0 - tw); };
-    const double eB0 = newEnter + teleEnter(newTeleFlow, newTeleWeight);
-    const double xB0 = newExit + teleExit(newTeleFlow, newTeleWeight);
-    const double fB0 = xB0 + newFlow;
-    const double eB1 = (newEnter + curEnter - deltaNew) + teleEnter(newTeleFlow + tfu, newTeleWeight + twu);
-    const double xB1 = (newExit + curExit - deltaNew) + teleExit(newTeleFlow + tfu, newTeleWeight + twu);
-    const double fB1 = xB1 + (newFlow + curFlow);
-
-    const double enterFlow1 = enterFlow + o.deltaEnterA + (eB1 - eB0);
-    const double d_enter = plogp(enterFlow1) - enterFlow_log_enterFlow;
-    const double d_enter_log = (o.plogpEA1 + plogp(eB1)) - (o.plogpEA0 + plogp(eB0));
-    const double d_exit_log = (o.plogpXA1 + plogp(xB1)) - (o.plogpXA0 + plogp(xB0));
-    const double d_flow_log = (o.plogpFA1 + plogp(fB1)) - (o.plogpFA0 + plogp(fB0));
-    return d_enter - d_enter_log - d_exit_log + d_flow_log;
-  }
-
-  // --- Non-redundant map equation (L*) per-module codebook terms ---------------
-  // Direct ports of MapEquation::nrEnterWithin / nrExitTerm. F is the module's
-  // sum_{leaf in module} plogp(leafFlow). Enter/exit are the module's boundary
-  // rates (separate for directed flow). See "Non-redundant map equation - expanded
-  // form.ipynb" for the derivation.
-  inline double nrEnterWithin(double moduleFlow, double qEnter, double qExit, double F)
-  {
-    using infomath::plogp;
-    if (moduleFlow < 1e-16)
-      return 0.0;
-    const double tEnter = qEnter * (plogp(moduleFlow) - F) / moduleFlow;
-    double tWithin = 0.0;
-    const double T = moduleFlow + qExit;
-    if (T > 1e-16) {
-      const double usage = moduleFlow + qExit - qEnter; // = moduleFlow when undirected
-      tWithin = (usage / T) * (plogp(T) - F - plogp(qExit));
-    }
-    return tEnter + tWithin;
-  }
-
-  // One module's leave-one-out exit codebook contribution, given the sibling
-  // enter-rate total-plus-exit-network sumEnterPlusE = (sum_b qEnter_b) + e and
-  // sumEnterLogEnter = sum_b plogp(qEnter_b); e is the exit-to-parent codeword.
-  inline double nrExitTerm(double qEnter, double qExit, double sumEnterPlusE, double sumEnterLogEnter, double e)
-  {
-    using infomath::plogp;
-    const double Z = sumEnterPlusE - qEnter;
-    if (Z < 1e-16)
-      return 0.0;
-    return qExit * (plogp(Z) - (sumEnterLogEnter - plogp(qEnter)) - plogp(e)) / Z;
-  }
-} // namespace
+using namespace columnar;
 
 void ColumnarTwoLevel::buildFromLeaves(const std::vector<InfoNode*>& leafNodes, bool undirected, unsigned long seed)
 {
@@ -2522,689 +1881,6 @@ double ColumnarTwoLevel::objectiveCorrection() const
   for (const auto& correction : m_corrections)
     sum += correction->hierarchicalCorrection(*this);
   return sum;
-}
-
-double BiasedEntropyCorrection::hierarchicalCorrection(const ColumnarTwoLevel& core) const
-{
-  // Sum of childDegree over all internal nodes incl. root == count of non-root
-  // nodes == sum of every level's size (leaves + modules at all levels).
-  long long nonRootNodes = 0;
-  const unsigned int levels = core.hierNumLevels();
-  for (unsigned int k = 0; k < levels; ++k)
-    nonRootNodes += core.hierLevelSize(static_cast<int>(k));
-  return m_multiplier * static_cast<double>(nonRootNodes) / (2.0 * m_totalDegree);
-}
-
-// ===================================================
-// PreferredModulesCorrection (--preferred-number-of-modules: |K - K_pref| bias)
-// ===================================================
-
-double PreferredModulesCorrection::hierarchicalCorrection(const ColumnarTwoLevel& core) const
-{
-  // Leaf-module level, matching where the move-loop bias acts (and the other
-  // corrections apply). <2 levels means no modules; the base handles that and
-  // the one-level fallback would collapse it anyway.
-  if (core.hierNumLevels() < 2)
-    return 0.0;
-  return cost(core.hierLevelSize(1));
-}
-
-double PreferredModulesCorrection::initMoveLoop(const std::vector<int>& leafModule, int numModules)
-{
-  m_moduleMembers.assign(numModules, 0);
-  for (int m : leafModule)
-    ++m_moduleMembers[m];
-  m_numNonEmpty = 0;
-  for (int c : m_moduleMembers)
-    if (c > 0)
-      ++m_numNonEmpty;
-  return cost(m_numNonEmpty);
-}
-
-double PreferredModulesCorrection::moveDelta(int /*leaf*/, int oldMod, int newMod) const
-{
-  if (oldMod == newMod)
-    return 0.0;
-  // K changes only when the move empties the old module or fills an empty new
-  // one (exactly BiasedMapEquation::getDeltaNumModulesIfMoving).
-  const int deltaK = (m_moduleMembers[oldMod] == 1 ? -1 : 0) + (m_moduleMembers[newMod] == 0 ? 1 : 0);
-  if (deltaK == 0)
-    return 0.0;
-  return cost(m_numNonEmpty + deltaK) - cost(m_numNonEmpty);
-}
-
-void PreferredModulesCorrection::applyMove(int /*leaf*/, int oldMod, int newMod)
-{
-  if (oldMod == newMod)
-    return;
-  if (m_moduleMembers[newMod] == 0)
-    ++m_numNonEmpty;
-  ++m_moduleMembers[newMod];
-  --m_moduleMembers[oldMod];
-  if (m_moduleMembers[oldMod] == 0)
-    --m_numNonEmpty;
-}
-
-double PreferredModulesCorrection::mergeDelta(int a, int b) const
-{
-  // Folding module A into B drops K by one only when both are non-empty.
-  const int deltaK = (m_moduleMembers[a] > 0 && m_moduleMembers[b] > 0) ? -1 : 0;
-  if (deltaK == 0)
-    return 0.0;
-  return cost(m_numNonEmpty + deltaK) - cost(m_numNonEmpty);
-}
-
-void PreferredModulesCorrection::applyMerge(int a, int b)
-{
-  if (m_moduleMembers[a] == 0)
-    return;
-  if (m_moduleMembers[b] == 0) {
-    // Non-empty A into empty B: B gains A's members as A empties — K unchanged.
-    m_moduleMembers[b] = m_moduleMembers[a];
-  } else {
-    m_moduleMembers[b] += m_moduleMembers[a];
-    --m_numNonEmpty;
-  }
-  m_moduleMembers[a] = 0;
-}
-
-double MetaCorrection::hierarchicalCorrection(const ColumnarTwoLevel& core) const
-{
-  using infomath::plogp;
-  // Meta term applies at the leaf-module level (level 1): for each bottom
-  // module, metaDataRate * F_m * (-sum_c plogp(f_c / F_m)) == metaDataRate *
-  // (plogp(F_m) - sum_c plogp(f_c)), matching MetaCollection::calculateEntropy
-  // summed over leaf modules.
-  if (core.hierNumLevels() < 2)
-    return 0.0;
-  const int nLeaves = core.numLeaves();
-  const int numModules = core.hierLevelSize(1);
-
-  std::vector<double> moduleFlow(numModules, 0.0);
-  std::unordered_map<long long, double> catFlow; // key = (module<<32)|category
-  for (int i = 0; i < nLeaves; ++i) {
-    const int m = core.hierLeafModule(i);
-    moduleFlow[m] += m_leafWeight[i];
-    const long long key = (static_cast<long long>(m) << 32) | static_cast<unsigned int>(m_leafCategory[i]);
-    catFlow[key] += m_leafWeight[i];
-  }
-
-  double total = 0.0;
-  for (int m = 0; m < numModules; ++m)
-    total += plogp(moduleFlow[m]);
-  for (const auto& kv : catFlow)
-    total -= plogp(kv.second);
-  return m_metaDataRate * total;
-}
-
-double MetaCorrection::moduleCategoryFlow(int module, int category) const
-{
-  const auto& cats = m_moduleCatFlow[module];
-  const auto it = cats.find(category);
-  return it == cats.end() ? 0.0 : it->second;
-}
-
-void MetaCorrection::setUnits(const std::vector<int>& leafToUnit, int numUnits)
-{
-  // Per-unit sparse (category -> weight) aggregates + total weight, ascending
-  // category id for deterministic iteration.
-  std::vector<std::unordered_map<int, double>> agg(numUnits);
-  m_unitWeight.assign(numUnits, 0.0);
-  const int nLeaves = static_cast<int>(leafToUnit.size());
-  for (int i = 0; i < nLeaves; ++i) {
-    agg[leafToUnit[i]][m_leafCategory[i]] += m_leafWeight[i];
-    m_unitWeight[leafToUnit[i]] += m_leafWeight[i];
-  }
-  m_unitCats.assign(numUnits, {});
-  for (int u = 0; u < numUnits; ++u) {
-    m_unitCats[u].assign(agg[u].begin(), agg[u].end());
-    std::sort(m_unitCats[u].begin(), m_unitCats[u].end());
-  }
-}
-
-void MetaCorrection::resetUnitsToLeaves()
-{
-  m_unitCats.clear();
-  m_unitWeight.clear();
-}
-
-double MetaCorrection::initMoveLoop(const std::vector<int>& unitModule, int numModules)
-{
-  using infomath::plogp;
-  m_moduleFlow.assign(numModules, 0.0);
-  m_moduleCatFlow.assign(numModules, {});
-  const int nUnits = static_cast<int>(unitModule.size());
-  if (m_unitCats.empty()) {
-    // Units are leaves.
-    for (int i = 0; i < nUnits; ++i) {
-      const int m = unitModule[i];
-      m_moduleFlow[m] += m_leafWeight[i];
-      m_moduleCatFlow[m][m_leafCategory[i]] += m_leafWeight[i];
-    }
-  } else {
-    for (int u = 0; u < nUnits; ++u) {
-      const int m = unitModule[u];
-      m_moduleFlow[m] += m_unitWeight[u];
-      for (const auto& cw : m_unitCats[u])
-        m_moduleCatFlow[m][cw.first] += cw.second;
-    }
-  }
-  double total = 0.0;
-  for (int m = 0; m < numModules; ++m)
-    total += plogp(m_moduleFlow[m]);
-  for (const auto& cats : m_moduleCatFlow)
-    for (const auto& kv : cats)
-      total -= plogp(kv.second);
-  return m_metaDataRate * total;
-}
-
-double MetaCorrection::moveDelta(int unit, int oldMod, int newMod) const
-{
-  using infomath::plogp;
-  if (oldMod == newMod)
-    return 0.0;
-  // term(m) = plogp(F_m) - sum_c plogp(f_c). Only F and the unit's categories
-  // change in the two modules.
-  const double Fo = m_moduleFlow[oldMod], Fn = m_moduleFlow[newMod];
-  if (m_unitCats.empty()) {
-    const int q = m_leafCategory[unit];
-    const double w = m_leafWeight[unit];
-    // The old-module delta is identical for every candidate the move loop probes
-    // for the same leaf, so it (and plogp(w)) is computed once and cached; a
-    // candidate module lacking category q (the common case) needs no category
-    // logs — plogp(0) == 0 and plogp(0 + w) == plogp(w), cached alongside.
-    if (unit != m_cacheUnit || oldMod != m_cacheOldMod) {
-      const double fo = moduleCategoryFlow(oldMod, q);
-      m_cacheUnit = unit;
-      m_cacheOldMod = oldMod;
-      m_cacheDOld = (plogp(Fo - w) - plogp(Fo)) - (plogp(fo - w) - plogp(fo));
-      m_cachePlogpW = plogp(w);
-    }
-    const double fn = moduleCategoryFlow(newMod, q);
-    const double dNew = (plogp(Fn + w) - plogp(Fn))
-        - (fn == 0.0 ? m_cachePlogpW : (plogp(fn + w) - plogp(fn)));
-    return m_metaDataRate * (m_cacheDOld + dNew);
-  }
-  const double w = m_unitWeight[unit];
-  double d = (plogp(Fo - w) - plogp(Fo)) + (plogp(Fn + w) - plogp(Fn));
-  for (const auto& cw : m_unitCats[unit]) {
-    const double fo = moduleCategoryFlow(oldMod, cw.first), fn = moduleCategoryFlow(newMod, cw.first);
-    d -= (plogp(fo - cw.second) - plogp(fo)) + (plogp(fn + cw.second) - plogp(fn));
-  }
-  return m_metaDataRate * d;
-}
-
-void MetaCorrection::applyMove(int unit, int oldMod, int newMod)
-{
-  if (oldMod == newMod)
-    return;
-  m_cacheUnit = -1; // module contents change: drop the per-leaf delta cache
-  auto moveWeight = [&](int q, double w) {
-    auto& oldCats = m_moduleCatFlow[oldMod];
-    auto oit = oldCats.find(q);
-    if (oit != oldCats.end()) {
-      oit->second -= w;
-      if (oit->second <= 1e-16)
-        oldCats.erase(oit);
-    }
-    m_moduleCatFlow[newMod][q] += w;
-  };
-  if (m_unitCats.empty()) {
-    const double w = m_leafWeight[unit];
-    m_moduleFlow[oldMod] -= w;
-    m_moduleFlow[newMod] += w;
-    moveWeight(m_leafCategory[unit], w);
-    return;
-  }
-  const double w = m_unitWeight[unit];
-  m_moduleFlow[oldMod] -= w;
-  m_moduleFlow[newMod] += w;
-  for (const auto& cw : m_unitCats[unit])
-    moveWeight(cw.first, cw.second);
-}
-
-double MetaCorrection::mergeDelta(int a, int b) const
-{
-  using infomath::plogp;
-  // contribution(m) = metaDataRate * (plogp(F_m) - sum_c plogp(f_{m,c})).
-  const double Fa = m_moduleFlow[a], Fb = m_moduleFlow[b];
-  double d = plogp(Fa + Fb) - plogp(Fa) - plogp(Fb); // change in the F term
-  const auto& ca = m_moduleCatFlow[a];
-  const auto& cb = m_moduleCatFlow[b];
-  for (const auto& kv : ca) {
-    const auto it = cb.find(kv.first);
-    if (it == cb.end())
-      continue; // category only in A: plogp(a+0) - plogp(a) - plogp(0) == 0
-    const double bv = it->second;
-    d -= plogp(kv.second + bv) - plogp(kv.second) - plogp(bv); // -sum_c plogp(f_c)
-  }
-  return m_metaDataRate * d;
-}
-
-void MetaCorrection::applyMerge(int a, int b)
-{
-  m_cacheUnit = -1;
-  m_moduleFlow[b] += m_moduleFlow[a];
-  m_moduleFlow[a] = 0.0;
-  auto& ca = m_moduleCatFlow[a];
-  auto& cb = m_moduleCatFlow[b];
-  for (const auto& kv : ca)
-    cb[kv.first] += kv.second;
-  ca.clear();
-}
-
-std::unique_ptr<ColumnarCorrection> MetaCorrection::sliceForLeaves(const std::vector<int>& globalLeafIds) const
-{
-  const int n = static_cast<int>(globalLeafIds.size());
-  std::vector<int> category(n);
-  std::vector<double> weight(n);
-  for (int j = 0; j < n; ++j) {
-    const int g = globalLeafIds[j];
-    category[j] = m_leafCategory[g];
-    weight[j] = m_leafWeight[g];
-  }
-  return std::make_unique<MetaCorrection>(std::move(category), std::move(weight), m_metaDataRate);
-}
-
-// ===================================================
-// MemCorrection (memory / state networks: physical-node codebook)
-// ===================================================
-
-MemCorrection::MemCorrection(std::vector<int> leafPhysical, std::vector<double> leafFlow)
-    : m_leafPhysical(std::move(leafPhysical)), m_leafFlow(std::move(leafFlow)), m_cState(0.0)
-{
-  using infomath::plogp;
-  for (double f : m_leafFlow)
-    m_cState += plogp(f);
-  // Compact the physical ids to [0, P): every downstream structure (module
-  // maps, reverse index) keys on the compact id, and the dense per-module
-  // lookup (see initMoveLoop) indexes with it directly.
-  std::unordered_map<int, int> compact;
-  compact.reserve(m_leafPhysical.size());
-  for (auto& p : m_leafPhysical) {
-    const auto r = compact.emplace(p, static_cast<int>(compact.size()));
-    p = r.first->second;
-  }
-  m_numPhys = static_cast<int>(compact.size());
-}
-
-double MemCorrection::physFlow(int module, int physical) const
-{
-  if (m_dense)
-    return m_densePhysFlow[static_cast<std::size_t>(module) * m_numPhys + physical];
-  const auto& pf = m_modulePhysFlow[module];
-  const auto it = pf.find(physical);
-  return it == pf.end() ? 0.0 : it->second;
-}
-
-double MemCorrection::hierarchicalCorrection(const ColumnarTwoLevel& core) const
-{
-  using infomath::plogp;
-  if (core.hierNumLevels() < 2)
-    return 0.0;
-  const int nLeaves = core.numLeaves();
-  std::unordered_map<long long, double> physFlowMap; // key = (module<<32)|physical
-  for (int i = 0; i < nLeaves; ++i) {
-    const int m = core.hierLeafModule(i);
-    const long long key = (static_cast<long long>(m) << 32) | static_cast<unsigned int>(m_leafPhysical[i]);
-    physFlowMap[key] += m_leafFlow[i];
-  }
-  double sum = 0.0;
-  for (const auto& kv : physFlowMap)
-    sum += plogp(kv.second);
-  return m_cState - sum;
-}
-
-void MemCorrection::setUnits(const std::vector<int>& leafToUnit, int numUnits)
-{
-  // Per-unit sparse (physical -> flow) aggregates, ascending physical id for
-  // deterministic iteration.
-  std::vector<std::unordered_map<int, double>> agg(numUnits);
-  const int nLeaves = static_cast<int>(leafToUnit.size());
-  for (int i = 0; i < nLeaves; ++i)
-    agg[leafToUnit[i]][m_leafPhysical[i]] += m_leafFlow[i];
-  m_unitPhys.assign(numUnits, {});
-  for (int u = 0; u < numUnits; ++u) {
-    m_unitPhys[u].assign(agg[u].begin(), agg[u].end());
-    std::sort(m_unitPhys[u].begin(), m_unitPhys[u].end());
-  }
-}
-
-void MemCorrection::resetUnitsToLeaves()
-{
-  m_unitPhys.clear();
-}
-
-double MemCorrection::initMoveLoop(const std::vector<int>& unitModule, int numModules)
-{
-  using infomath::plogp;
-  m_modulePhysFlow.assign(numModules, {});
-  m_physModules.clear();
-  m_cacheUnit = -1;
-  // Dense per-module physical-flow lookup when it fits comfortably in memory
-  // (state networks typically have far fewer physical nodes than state nodes):
-  // physFlow becomes an O(1) array read in the move loop's hot path instead of
-  // a hash find. The sparse maps are kept alongside for iteration (merge scan).
-  m_dense = static_cast<double>(numModules) * static_cast<double>(m_numPhys) <= 8e6;
-  if (m_dense)
-    m_densePhysFlow.assign(static_cast<std::size_t>(numModules) * m_numPhys, 0.0);
-  // The physical -> modules reverse index only feeds the co-physical proposals
-  // (coMergeMode); maintaining it is pure overhead when that tuning mode is off.
-  const bool maintainIndex = coMergeMode() != 0;
-  const int nUnits = static_cast<int>(unitModule.size());
-  if (m_unitPhys.empty()) {
-    // Units are leaves.
-    for (int i = 0; i < nUnits; ++i) {
-      const int m = unitModule[i], p = m_leafPhysical[i];
-      m_modulePhysFlow[m][p] += m_leafFlow[i];
-      if (m_dense)
-        m_densePhysFlow[static_cast<std::size_t>(m) * m_numPhys + p] += m_leafFlow[i];
-      if (maintainIndex)
-        m_physModules[p].insert(m);
-    }
-  } else {
-    for (int u = 0; u < nUnits; ++u) {
-      const int m = unitModule[u];
-      for (const auto& pf : m_unitPhys[u]) {
-        m_modulePhysFlow[m][pf.first] += pf.second;
-        if (m_dense)
-          m_densePhysFlow[static_cast<std::size_t>(m) * m_numPhys + pf.first] += pf.second;
-        if (maintainIndex)
-          m_physModules[pf.first].insert(m);
-      }
-    }
-  }
-  double sum = 0.0;
-  for (const auto& pf : m_modulePhysFlow)
-    for (const auto& kv : pf)
-      sum += plogp(kv.second);
-  return m_cState - sum;
-}
-
-double MemCorrection::moveDelta(int unit, int oldMod, int newMod) const
-{
-  using infomath::plogp;
-  if (oldMod == newMod)
-    return 0.0;
-  // term = C_state - sum plogp(physFlow); only the unit's physicals in the two
-  // modules change. The old-module part is identical for every candidate of the
-  // same unit, so it is computed once and cached (the move loop probes many
-  // candidates per unit); a candidate without the unit's physical (the common
-  // case) needs no logs at all: plogp(0) == 0 and plogp(0 + f) == plogp(f),
-  // which is cached alongside.
-  if (m_unitPhys.empty()) {
-    const int p = m_leafPhysical[unit];
-    const double f = m_leafFlow[unit];
-    if (unit != m_cacheUnit || oldMod != m_cacheOldMod) {
-      const double oc = physFlow(oldMod, p);
-      m_cacheUnit = unit;
-      m_cacheOldMod = oldMod;
-      m_cacheOldTerm = plogp(oc) - plogp(oc - f);
-      m_cachePlogpF = plogp(f);
-    }
-    const double nc = physFlow(newMod, p);
-    return m_cacheOldTerm + (nc == 0.0 ? -m_cachePlogpF : (plogp(nc) - plogp(nc + f)));
-  }
-  double d = 0.0;
-  for (const auto& pf : m_unitPhys[unit]) {
-    const double oc = physFlow(oldMod, pf.first), nc = physFlow(newMod, pf.first);
-    const double dOld = (oc == 0.0) ? 0.0 : (plogp(oc) - plogp(oc - pf.second));
-    const double dNew = (nc == 0.0) ? -plogp(pf.second) : (plogp(nc) - plogp(nc + pf.second));
-    d += dOld + dNew;
-  }
-  return d;
-}
-
-void MemCorrection::applyMove(int unit, int oldMod, int newMod)
-{
-  if (oldMod == newMod)
-    return;
-  m_cacheUnit = -1; // module contents change: drop the per-unit delta cache
-  const bool maintainIndex = !m_physModules.empty() || coMergeMode() != 0;
-  auto moveFlow = [&](int p, double f) {
-    auto& oldPhys = m_modulePhysFlow[oldMod];
-    auto it = oldPhys.find(p);
-    if (it != oldPhys.end()) {
-      it->second -= f;
-      if (it->second <= 1e-16) {
-        oldPhys.erase(it);
-        if (maintainIndex)
-          m_physModules[p].erase(oldMod); // oldMod no longer holds physical p
-      }
-    }
-    m_modulePhysFlow[newMod][p] += f;
-    if (m_dense) {
-      auto& ov = m_densePhysFlow[static_cast<std::size_t>(oldMod) * m_numPhys + p];
-      ov -= f;
-      if (ov <= 1e-16)
-        ov = 0.0; // keep the zero fast path exact
-      m_densePhysFlow[static_cast<std::size_t>(newMod) * m_numPhys + p] += f;
-    }
-    if (maintainIndex)
-      m_physModules[p].insert(newMod);
-  };
-  if (m_unitPhys.empty()) {
-    moveFlow(m_leafPhysical[unit], m_leafFlow[unit]);
-    return;
-  }
-  for (const auto& pf : m_unitPhys[unit])
-    moveFlow(pf.first, pf.second);
-}
-
-void MemCorrection::proposeMoveTargets(int unit, std::vector<int>& out) const
-{
-  // Modules already holding a state node of this unit's physical node(s) — the
-  // co-physical merges that shrink the physical codebook but are (usually) not
-  // reachable through the edge neighborhood.
-  if (m_unitPhys.empty()) {
-    const auto it = m_physModules.find(m_leafPhysical[unit]);
-    if (it == m_physModules.end())
-      return;
-    for (int m : it->second)
-      out.push_back(m);
-    return;
-  }
-  const auto base = static_cast<std::ptrdiff_t>(out.size());
-  for (const auto& pf : m_unitPhys[unit]) {
-    const auto it = m_physModules.find(pf.first);
-    if (it == m_physModules.end())
-      continue;
-    for (int m : it->second)
-      out.push_back(m);
-  }
-  std::sort(out.begin() + base, out.end());
-  out.erase(std::unique(out.begin() + base, out.end()), out.end());
-}
-
-double MemCorrection::mergeDelta(int a, int b) const
-{
-  using infomath::plogp;
-  // contribution(m) = -sum_p plogp(physFlow_m[p]); merging folds A's physical
-  // flows into B. Only physicals present in BOTH modules change the sum (a
-  // physical in only one keeps its single plogp term). Iterate the (smaller)
-  // A side; b-only physicals are handled when B is the A of another pair.
-  const auto& pa = m_modulePhysFlow[a];
-  double dSum = 0.0; // change in sum_p plogp(physFlow)
-  if (m_dense) {
-    const std::size_t bBase = static_cast<std::size_t>(b) * m_numPhys;
-    for (const auto& kv : pa) {
-      const double bv = m_densePhysFlow[bBase + kv.first];
-      if (bv == 0.0)
-        continue; // physical only in A: plogp(a+0) - plogp(a) - plogp(0) == 0
-      dSum += plogp(kv.second + bv) - plogp(kv.second) - plogp(bv);
-    }
-  } else {
-    const auto& pb = m_modulePhysFlow[b];
-    for (const auto& kv : pa) {
-      const auto it = pb.find(kv.first);
-      if (it == pb.end())
-        continue; // physical only in A: plogp(a+0) - plogp(a) - plogp(0) == 0
-      dSum += plogp(kv.second + it->second) - plogp(kv.second) - plogp(it->second);
-    }
-  }
-  return -dSum; // correction = C_state - sum_p plogp; delta correction = -dSum
-}
-
-void MemCorrection::applyMerge(int a, int b)
-{
-  m_cacheUnit = -1;
-  auto& pa = m_modulePhysFlow[a];
-  auto& pb = m_modulePhysFlow[b];
-  for (const auto& kv : pa) {
-    pb[kv.first] += kv.second;
-    if (m_dense) {
-      m_densePhysFlow[static_cast<std::size_t>(b) * m_numPhys + kv.first] += kv.second;
-      m_densePhysFlow[static_cast<std::size_t>(a) * m_numPhys + kv.first] = 0.0;
-    }
-    if (!m_physModules.empty()) { // maintained only when co-physical mode is on
-      auto& mods = m_physModules[kv.first];
-      mods.erase(a);
-      mods.insert(b);
-    }
-  }
-  pa.clear();
-}
-
-void MemCorrection::proposeMergePartners(int module, std::vector<int>& out) const
-{
-  // Leaf modules sharing a physical node with `module` (co-physical merge
-  // candidates the edge-based set can miss on directed clustering).
-  if (module >= static_cast<int>(m_modulePhysFlow.size()))
-    return;
-  for (const auto& kv : m_modulePhysFlow[module]) {
-    const auto it = m_physModules.find(kv.first);
-    if (it == m_physModules.end())
-      continue;
-    for (int m : it->second)
-      if (m != module)
-        out.push_back(m);
-  }
-}
-
-std::unique_ptr<ColumnarCorrection> MemCorrection::sliceForLeaves(const std::vector<int>& globalLeafIds) const
-{
-  const int n = static_cast<int>(globalLeafIds.size());
-  std::vector<int> physical(n);
-  std::vector<double> flow(n);
-  for (int j = 0; j < n; ++j) {
-    const int g = globalLeafIds[j];
-    physical[j] = m_leafPhysical[g];
-    flow[j] = m_leafFlow[g];
-  }
-  return std::make_unique<MemCorrection>(std::move(physical), std::move(flow));
-}
-
-// ===================================================
-// LossyCorrection (rate-distortion map equation: noise modules)
-// ===================================================
-
-LossyCorrection::LossyCorrection(std::vector<double> leafFlow, std::vector<double> leafEntropy, double lambda)
-    : m_leafFlow(std::move(leafFlow)), m_leafEntropy(std::move(leafEntropy)), m_lambda(lambda)
-{
-  using infomath::plogp;
-  m_leafFlf.resize(m_leafFlow.size());
-  for (std::size_t i = 0; i < m_leafFlow.size(); ++i)
-    m_leafFlf[i] = plogp(m_leafFlow[i]);
-}
-
-double LossyCorrection::moduleCost(double flow, double flowLogFlow, double entropy) const
-{
-  using infomath::plogp;
-  // Naming-overhead loss beyond the tolerated distortion; 0 when the module is
-  // not noise (matches LossyMapEquation::calcCorrection).
-  return std::max(0.0, (plogp(flow) - flowLogFlow) - m_lambda * entropy);
-}
-
-double LossyCorrection::hierarchicalCorrection(const ColumnarTwoLevel& core) const
-{
-  if (core.hierNumLevels() < 2)
-    return 0.0;
-  const int nLeaves = core.numLeaves();
-  const int numModules = core.hierLevelSize(1);
-  std::vector<double> F(numModules, 0.0), flf(numModules, 0.0), H(numModules, 0.0);
-  for (int i = 0; i < nLeaves; ++i) {
-    const int m = core.hierLeafModule(i);
-    F[m] += m_leafFlow[i];
-    flf[m] += m_leafFlf[i];
-    H[m] += m_leafEntropy[i];
-  }
-  double sumC = 0.0;
-  for (int m = 0; m < numModules; ++m)
-    sumC += moduleCost(F[m], flf[m], H[m]);
-  return -sumC; // objective J = base - sum_m c_m
-}
-
-double LossyCorrection::initMoveLoop(const std::vector<int>& leafModule, int numModules)
-{
-  m_moduleFlow.assign(numModules, 0.0);
-  m_moduleFlf.assign(numModules, 0.0);
-  m_moduleEntropy.assign(numModules, 0.0);
-  const int nLeaves = static_cast<int>(leafModule.size());
-  for (int i = 0; i < nLeaves; ++i) {
-    const int m = leafModule[i];
-    m_moduleFlow[m] += m_leafFlow[i];
-    m_moduleFlf[m] += m_leafFlf[i];
-    m_moduleEntropy[m] += m_leafEntropy[i];
-  }
-  double sumC = 0.0;
-  for (int m = 0; m < numModules; ++m)
-    sumC += moduleCost(m_moduleFlow[m], m_moduleFlf[m], m_moduleEntropy[m]);
-  return -sumC;
-}
-
-double LossyCorrection::moveDelta(int leaf, int oldMod, int newMod) const
-{
-  if (oldMod == newMod)
-    return 0.0;
-  const double f = m_leafFlow[leaf], lf = m_leafFlf[leaf], h = m_leafEntropy[leaf];
-  const double dOld = moduleCost(m_moduleFlow[oldMod] - f, m_moduleFlf[oldMod] - lf, m_moduleEntropy[oldMod] - h)
-      - moduleCost(m_moduleFlow[oldMod], m_moduleFlf[oldMod], m_moduleEntropy[oldMod]);
-  const double dNew = moduleCost(m_moduleFlow[newMod] + f, m_moduleFlf[newMod] + lf, m_moduleEntropy[newMod] + h)
-      - moduleCost(m_moduleFlow[newMod], m_moduleFlf[newMod], m_moduleEntropy[newMod]);
-  return -(dOld + dNew); // correction = -sum_m c_m
-}
-
-void LossyCorrection::applyMove(int leaf, int oldMod, int newMod)
-{
-  if (oldMod == newMod)
-    return;
-  const double f = m_leafFlow[leaf], lf = m_leafFlf[leaf], h = m_leafEntropy[leaf];
-  m_moduleFlow[oldMod] -= f;
-  m_moduleFlf[oldMod] -= lf;
-  m_moduleEntropy[oldMod] -= h;
-  m_moduleFlow[newMod] += f;
-  m_moduleFlf[newMod] += lf;
-  m_moduleEntropy[newMod] += h;
-}
-
-double LossyCorrection::mergeDelta(int a, int b) const
-{
-  const double ca = moduleCost(m_moduleFlow[a], m_moduleFlf[a], m_moduleEntropy[a]);
-  const double cb = moduleCost(m_moduleFlow[b], m_moduleFlf[b], m_moduleEntropy[b]);
-  const double cab = moduleCost(m_moduleFlow[a] + m_moduleFlow[b],
-                                m_moduleFlf[a] + m_moduleFlf[b],
-                                m_moduleEntropy[a] + m_moduleEntropy[b]);
-  return -(cab - ca - cb); // change in -sum_m c_m
-}
-
-void LossyCorrection::applyMerge(int a, int b)
-{
-  m_moduleFlow[b] += m_moduleFlow[a];
-  m_moduleFlf[b] += m_moduleFlf[a];
-  m_moduleEntropy[b] += m_moduleEntropy[a];
-  m_moduleFlow[a] = m_moduleFlf[a] = m_moduleEntropy[a] = 0.0;
-}
-
-std::unique_ptr<ColumnarCorrection> LossyCorrection::sliceForLeaves(const std::vector<int>& globalLeafIds) const
-{
-  const int n = static_cast<int>(globalLeafIds.size());
-  std::vector<double> flow(n), entropy(n);
-  for (int j = 0; j < n; ++j) {
-    const int g = globalLeafIds[j];
-    flow[j] = m_leafFlow[g];
-    entropy[j] = m_leafEntropy[g];
-  }
-  return std::make_unique<LossyCorrection>(std::move(flow), std::move(entropy), m_lambda);
 }
 
 void ColumnarTwoLevel::addSlicedLeafCorrections(ColumnarTwoLevel& subOpt, const std::vector<int>& globalLeafIds) const
